Update session.rb
Creates the 'start' method, available to ActionDispatch::Request::Session objects.  This method allows multiple session variables to be initialized in a single line of code.  Default values can be assigned to session keys when an application starts, but are overridden by other methods (similar to the ||= operator).  Practical uses include form data persistence where the method rendering the form can populate each field with session variables, and then modify them if the form is submitted multiple times (e.g. because of invalid data).  If the session has been 'start'ed with default values, these will appear in the form fields, and the application will not raise an error that the requested keys do not exist.